# 8352075: Perf regression accessing fields