--- a/src/hotspot/share/oops/fieldStreams.hpp
+++ b/src/hotspot/share/oops/fieldStreams.hpp
@@ -42,29 +42,44 @@ class FieldStreamBase : public StackObj {
  protected:
   const Array<u1>*    _fieldinfo_stream;
   FieldInfoReader     _reader;
+  int                 _ctrl_offset;
   constantPoolHandle  _constants;
   int                 _index;
   int                 _limit;
 
   FieldInfo           _fi_buf;
   fieldDescriptor     _fd_buf;
+  int                 _next_field_offset;
 
-  FieldInfo const * field() const {
+  void ensure_info() {
     assert(!done(), "no more fields");
-    return &_fi_buf;
+    if (_next_field_offset > _reader.position()) {
+      _reader.read_partial_record(_fi_buf);
+      _next_field_offset = _reader.position();
+    }
   }
 
   inline FieldStreamBase(const Array<u1>* fieldinfo_stream, ConstantPool* constants, int start, int limit);
 
-  inline FieldStreamBase(Array<u1>* fieldinfo_stream, ConstantPool* constants);
+  inline FieldStreamBase(const Array<u1>* fieldinfo_stream, ConstantPool* constants);
 
-  private:
+ private:
+  void read_next() {
+    int ctrl_byte = _fieldinfo_stream->at(_ctrl_offset + _index);
+    _next_field_offset += ctrl_byte & CTRL_LENGTH_MASK;
+    _reader.read_name_signature(_fi_buf);
+    _fi_buf.field_flags_addr()->update_injected(ctrl_byte & INJECTED_FIELD);
+  }
+
    void initialize() {
     int java_fields_count = _reader.next_uint();
     int injected_fields_count = _reader.next_uint();
+    _ctrl_offset = _reader.position();
+    _reader.skip_bytes(java_fields_count + injected_fields_count);
     assert( _limit <= java_fields_count + injected_fields_count, "Safety check");
     if (_limit != 0) {
-      _reader.read_field_info(_fi_buf);
+      _next_field_offset = _reader.position();
+      read_next();
     }
    }
  public:
@@ -76,51 +91,58 @@ class FieldStreamBase : public StackObj {
 
   void next() {
     _index += 1;
+    _reader.set_position_and_next_index(_next_field_offset, _index);
     if (done()) return;
-    _reader.read_field_info(_fi_buf);
+    read_next();
   }
   bool done() const { return _index >= _limit; }
 
   // Accessors for current field
-  AccessFlags access_flags() const {
-    return field()->access_flags();
+  AccessFlags access_flags() {
+    ensure_info();
+    return _fi_buf.access_flags();
   }
 
-  FieldInfo::FieldFlags field_flags() const {
-    return field()->field_flags();
+  FieldInfo::FieldFlags field_flags() {
+    ensure_info();
+    return _fi_buf.field_flags();
   }
 
   Symbol* name() const {
-    return field()->name(_constants());
+    return _fi_buf.name(_constants());
   }
 
   Symbol* signature() const {
-    return field()->signature(_constants());
+    return _fi_buf.signature(_constants());
   }
 
-  Symbol* generic_signature() const {
-    if (field()->field_flags().is_generic()) {
-      return _constants->symbol_at(field()->generic_signature_index());
+  Symbol* generic_signature() {
+    if (_fi_buf.field_flags().is_generic()) {
+      return _constants->symbol_at(_fi_buf.generic_signature_index());
     } else {
       return nullptr;
     }
   }
 
-  int offset() const {
-    return field()->offset();
+  int offset() {
+    ensure_info();
+    return _fi_buf.offset();
   }
 
   bool is_contended() const {
-    return field()->is_contended();
+    assert(_next_field_offset == _reader.position(), "unexpected");
+    return _fi_buf.is_contended();
   }
 
   int contended_group() const {
-    return field()->contended_group();
+    assert(_next_field_offset == _reader.position(), "unexpected");
+    return _fi_buf.contended_group();
   }
 
   // Convenient methods
 
-  const FieldInfo& to_FieldInfo() const {
+  const FieldInfo& to_FieldInfo() {
+    ensure_info();
     return _fi_buf;
   }
 
@@ -129,7 +151,8 @@ class FieldStreamBase : public StackObj {
   }
 
   // bridge to a heavier API:
-  fieldDescriptor& field_descriptor() const {
+  fieldDescriptor& field_descriptor() {
+    ensure_info();
     fieldDescriptor& field = const_cast<fieldDescriptor&>(_fd_buf);
     field.reinitialize(field_holder(), to_FieldInfo());
     return field;
@@ -142,27 +165,26 @@ class JavaFieldStream : public FieldStreamBase {
   JavaFieldStream(const InstanceKlass* k): FieldStreamBase(k->fieldinfo_stream(), k->constants(), 0, k->java_fields_count()) {}
 
   u2 name_index() const {
-    assert(!field()->field_flags().is_injected(), "regular only");
-    return field()->name_index();
+    return _fi_buf.name_index();
   }
 
   u2 signature_index() const {
-    assert(!field()->field_flags().is_injected(), "regular only");
-    return field()->signature_index();
-    return -1;
+    return _fi_buf.signature_index();
   }
 
-  u2 generic_signature_index() const {
-    assert(!field()->field_flags().is_injected(), "regular only");
-    if (field()->field_flags().is_generic()) {
-      return field()->generic_signature_index();
+  u2 generic_signature_index() {
+    ensure_info();
+    assert(!_fi_buf.field_flags().is_injected(), "regular only");
+    if (_fi_buf.field_flags().is_generic()) {
+      return _fi_buf.generic_signature_index();
     }
     return 0;
   }
 
-  u2 initval_index() const {
-    assert(!field()->field_flags().is_injected(), "regular only");
-    return field()->initializer_index();
+  u2 initval_index() {
+    ensure_info();
+    assert(!_fi_buf.field_flags().is_injected(), "regular only");
+    return _fi_buf.initializer_index();
   }
 };
 
@@ -176,7 +198,6 @@ class InternalFieldStream : public FieldStreamBase {
 
 class AllFieldStream : public FieldStreamBase {
  public:
-  AllFieldStream(Array<u1>* fieldinfo, ConstantPool* constants): FieldStreamBase(fieldinfo, constants) {}
   AllFieldStream(const InstanceKlass* k):      FieldStreamBase(k->fieldinfo_stream(), k->constants()) {}
 };
 
@@ -238,11 +259,11 @@ class HierarchicalFieldStream : public StackObj  {
 
   // bridge functions from FieldStreamBase
 
-  AccessFlags access_flags() const {
+  AccessFlags access_flags() {
     return _current_stream.access_flags();
   }
 
-  FieldInfo::FieldFlags field_flags() const {
+  FieldInfo::FieldFlags field_flags() {
     return _current_stream.field_flags();
   }
 
@@ -254,11 +275,11 @@ class HierarchicalFieldStream : public StackObj  {
     return _current_stream.signature();
   }
 
-  Symbol* generic_signature() const {
+  Symbol* generic_signature() {
     return _current_stream.generic_signature();
   }
 
-  int offset() const {
+  int offset() {
     return _current_stream.offset();
   }
 
@@ -274,7 +295,7 @@ class HierarchicalFieldStream : public StackObj  {
     return _current_stream.to_FieldInfo();
   }
 
-  fieldDescriptor& field_descriptor() const {
+  fieldDescriptor& field_descriptor() {
     return _current_stream.field_descriptor();
   }