--- a/src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/code/CompressedStream.java
+++ b/src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/code/CompressedStream.java
@@ -63,6 +63,10 @@ public class CompressedStream {
     this.position = position;
   }
 
+  public void skipBytes(int bytes) {
+    this.position += bytes;
+  }
+
   public int encodeSign(int value) {
     return Unsigned5.encodeSign(value);
   }