--- a/src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/oops/Field.java
+++ b/src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/oops/Field.java
@@ -111,6 +111,7 @@ public class Field {
     int numJavaFields = crs.readInt();     // read num_java_fields
     int numInjectedFields = crs.readInt(); // read num_injected_fields;
     int numFields = numJavaFields + numInjectedFields;
+    crs.skipBytes(numFields);
     Field[] fields = new Field[numFields];
     for (int i = 0; i < numFields; i++) {
       FieldInfoValues values = readFieldInfoValues(crs);